README.md
As we have planned to move the code to master branch only when we release versions of software, we now have all our code in develop. As, the Code climate is integrated with master branch, the status of code climate shows a '?'. As per Code climate documentation, we cannot change the default branch that was connected during integration and we need to request their support team to do it at their back end. we are waiting for their response.